requirements.txt
watchdog>=3.0.0
sentence-transformers>=2.2.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
python-pptx>=0.6.21
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
